+++ greenstep/src/main/java/com/mm/greenstep/domain/plogging/dto/response/PloggingDetailResDto.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Getter
 public class PloggingDetailResDto {
     private Integer getExp;
-    private Double travelTime;
+    private Long travelTime;
     private Double travelRange;
     private Integer trashAmount;
     private LocalDateTime createdAt;
