+++ greenstep/src/main/java/com/mm/greenstep/domain/plogging/entity/Plogging.java
@@ -30,7 +30,7 @@ public class Plogging {
     private Double travelRange;
 
     @Column(name = "travel_time")
-    private Double travelTime;
+    private Long travelTime;
 
     @Column(name = "travel_amount")
     private Integer trashAmount;
